String

+++ src/chap03/textbook/Exercise.java
@@ -5,7 +5,7 @@ public class Exercise {
 		int lengthTop = 5;
 		int lengthBottom = 10;
 		int height = 7;
-		double area =  ((lengthTop + lengthBottom) * height / 2);
+		double area =  ((double) (lengthTop + lengthBottom) * height / 2);
 				System.out.println(area);
 				}
